new endpoints added (get close item, get random item)
* config file updated to have different keys for server, to have
mappings between database and server keys, new keys to pay attention to
while analyzing data. other files are updated to use the new keys
* while processing csv files, in the flat version an object that holds
id's of rows as keys is produced instead of an array. other files are
updated to use this structure
* 2 routes added to routes for getting a random item and item that is
close to a given one.
* these functionalities are also added to controllers/analysis and
modules/analysis files
* a util file is added, it has only one function to check if a variable
holds a valid data value, more functions can be added in the future